+++ packages/react/src/components/Button/tokens.ts
@@ -4,7 +4,7 @@ export const tokens = (defaults: DefaultTokens) => {
   return {
     "border-radius": "2px",
     height: "48px",
-    "font-size": defaults.theme.typo.l,
-    "font-weight": defaults.theme.typo.medium,
+    "font-size": defaults.theme.font.sizes.l,
+    "font-weight": defaults.theme.font.weights.medium,
   };
 };