+++ packages/tokens/src/bin/__mocks__/cunningham.ts
@@ -4,5 +4,19 @@ module.exports = {
       primary: "#055FD2",
       secondary: "#DA0000",
     },
+    font: {
+      sizes: {
+        m: "1rem",
+      },
+      weights: {
+        medium: 400,
+      },
+      families: {
+        base: "Roboto",
+      },
+    },
+    spacings: {
+      s: "1rem",
+    },
   },
 };